GOB: move inter code into seperate folder for inter
move all inter related source code files into an folder for better code overview, like for save or sound
