+++ engines/gob/console.cpp
@@ -27,7 +27,7 @@
 
 #include "gob/console.h"
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/dataio.h"
 #include "gob/cheater.h"
