+++ engines/gob/demos/demoplayer.cpp
@@ -34,7 +34,7 @@
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/draw.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
