+++ engines/gob/demos/scnplayer.cpp
@@ -32,7 +32,7 @@
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/draw.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/video/videoplayer.h"
 
 namespace Gob {
