+++ engines/gob/draw.cpp
@@ -36,7 +36,7 @@
 #include "gob/game.h"
 #include "gob/resources.h"
 #include "gob/script.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/video/video.h"
 #include "gob/palanim.h"
