+++ engines/gob/draw_playtoons.cpp
@@ -28,7 +28,7 @@
 #include "common/endian.h"
 
 #include "gob/draw.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/game.h"
 #include "gob/resources.h"
