+++ engines/gob/game.cpp
@@ -36,7 +36,7 @@
 #include "gob/script.h"
 #include "gob/resources.h"
 #include "gob/hotspots.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/draw.h"
 #include "gob/mult.h"
 #include "gob/scenery.h"
