+++ engines/gob/gob.cpp
@@ -45,7 +45,7 @@
 #include "gob/game.h"
 #include "gob/sound/sound.h"
 #include "gob/init.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/draw.h"
 #include "gob/goblin.h"
 #include "gob/map.h"
