+++ engines/gob/goblin.cpp
@@ -36,7 +36,7 @@
 #include "gob/map.h"
 #include "gob/mult.h"
 #include "gob/scenery.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/sound/sound.h"
 
 namespace Gob {
