+++ engines/gob/goblin_v2.cpp
@@ -35,7 +35,7 @@
 #include "gob/map.h"
 #include "gob/mult.h"
 #include "gob/scenery.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 
 namespace Gob {
