+++ engines/gob/goblin_v4.cpp
@@ -31,7 +31,7 @@
 #include "gob/mult.h"
 #include "gob/map.h"
 #include "gob/scenery.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 
 namespace Gob {
