+++ engines/gob/hotspots.cpp
@@ -32,7 +32,7 @@
 #include "gob/draw.h"
 #include "gob/game.h"
 #include "gob/script.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 
 namespace Gob {
