+++ engines/gob/inter/inter_fascin.cpp
@@ -29,7 +29,7 @@
 
 #include "gob/hotspots.h"
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/dataio.h"
