+++ engines/gob/inter/inter_littlered.cpp
@@ -26,7 +26,7 @@
  */
 
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/draw.h"
