+++ engines/gob/inter/inter_v1.cpp
@@ -28,7 +28,7 @@
 #include "common/str.h"
 
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/dataio.h"
@@ -39,7 +39,6 @@
 #include "gob/resources.h"
 #include "gob/hotspots.h"
 #include "gob/goblin.h"
-#include "gob/inter.h"
 #include "gob/map.h"
 #include "gob/mult.h"
 #include "gob/palanim.h"
