+++ engines/gob/inter/inter_v2.cpp
@@ -32,7 +32,7 @@
 #include "gui/message.h"
 
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/dataio.h"
