+++ engines/gob/inter/inter_v3.cpp
@@ -29,7 +29,7 @@
 #include "common/file.h"
 
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/dataio.h"
 #include "gob/draw.h"
