+++ engines/gob/inter/inter_v5.cpp
@@ -30,7 +30,7 @@
 #include "gui/message.h"
 
 #include "gob/gob.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/global.h"
 #include "gob/game.h"
 #include "gob/script.h"
