+++ engines/gob/save/saveload_v4.cpp
@@ -27,7 +27,7 @@
 
 #include "gob/save/saveload.h"
 #include "gob/save/saveconverter.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/variables.h"
 
 namespace Gob {
