+++ engines/gob/scenery.cpp
@@ -35,7 +35,7 @@
 #include "gob/game.h"
 #include "gob/script.h"
 #include "gob/resources.h"
-#include "gob/inter.h"
+#include "gob/inter/inter.h"
 #include "gob/map.h"
 #include "gob/video/videoplayer.h"
